refactor: splitted comp into page specific,shared, added umami npm package,moved umami credential to .env
