+++ package.json
@@ -15,6 +15,7 @@
   },
   "dependencies": {
     "next": "15.1.3",
+    "next-umami": "^0.0.4",
     "phosphor-react": "^1.4.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
