+++ src/app/page.tsx
@@ -1,22 +1,12 @@
-import Script from 'next/script';
-import CallToAction from './calltoaction';
-import Hero from './hero';
-import NavBar from './nav';
-import Events from './events';
+import CallToAction from '../components/pages/home/calltoaction';
+import Hero from '../components/pages/home/hero';
+import Events from '../components/pages/home/events';
 
 export default function Home() {
   return (
     <>
-      <Script
-        src='https://cloud.umami.is/script.js'
-        data-website-id='ff99f01a-248c-4f31-87f7-fc40499f219a'
-        strategy='afterInteractive'
-        async
-        defer
-      />
-      <div className='bg-[#fafafa]'>
+      <div>
         <div className='mx-auto max-w-[1120px] font-inter'>
-          <NavBar />
           <Hero />
           <Events />
           <CallToAction />
