+++ src/components/shared/footer.tsx
@@ -1,0 +1,12 @@
+export default function Footer() {
+  return (
+    <footer className='flex w-full flex-1 items-center justify-center bg-[#fafafa] px-4 py-4 text-black md:px-8 lg:px-16'>
+      <p className='mb-2 mt-16 text-center text-gray-600'>
+        Made with luv from Hari and Justin 💚
+        <a href='https://fossunited.org/c/chennai' className='ml-2 text-[#03b051]'>
+          Foss United Chennai
+        </a>
+      </p>
+    </footer>
+  );
+}
