+++ src/components/shared/header.tsx
@@ -1,16 +1,15 @@
-import GitHubButton from '../components/GitHubButton';
+import Link from 'next/link';
+import GitHubButton from '../github-button';
 
-const NavBar = () => {
+export default function Header() {
   return (
     <nav className='flex w-full flex-1 items-center justify-between bg-[#fafafa] px-4 py-4 text-black md:px-8 lg:px-16'>
-      <a href='#' className='text-xl font-semibold'>
+      <Link href='/' className='text-xl font-semibold'>
         tamilnadu.tech
-      </a>
+      </Link>
       <div className='space-x-4'>
         <GitHubButton />
       </div>
     </nav>
   );
-};
-
-export default NavBar;
+}
