+++ src/lib/constants.ts
@@ -1,3 +1,5 @@
 const SITE_URL = 'https://tamilnadu.tech';
 
-export { SITE_URL };
+const IS_PROD = process.env.NODE_ENV === 'production';
+
+export { SITE_URL, IS_PROD };
